requirements.txt
# Python dependencies for Asset Processing
mysql-connector-python>=8.0.33
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
xlwt>=1.3.0
xlutils>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
msoffcrypto-tool>=5.0.0
